Move tensor block model to blockmodel.py and break out orientation metadata
